+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: hakader <hakader@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 16:05:52 by hakader           #+#    #+#             */
-/*   Updated: 2024/11/05 22:56:25 by hakader          ###   ########.fr       */
+/*   Updated: 2024/11/08 21:56:11 by hakader          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strdup(const char *src)
 {
-	size_t	size;
+	int		size;
 	char	*dest;
 
 	size = ft_strlen(src);
